fix: rename Patient controller for BackendPatient microservice and configure projects folders to fix file issues
