--- BackendPatient/Controllers/Patient.cs
@@ -1,95 +1,0 @@
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-using BackendPatient.Data;
-using BackendPatient.Models;
-
-namespace BackendPatient.Controllers;
-
-[ApiController]
-[Route("api/[controller]")]
-public class PatientsController : ControllerBase
-{
-    private readonly ApplicationDbContext _dbContext;
-
-    public PatientsController(ApplicationDbContext dbContext)
-    {
-        _dbContext = dbContext;
-    }
-
-    [HttpGet]
-    public async Task<ActionResult<IEnumerable<Patient>>> GetPatients()
-    {
-        return await _dbContext.Patients.ToListAsync();
-    }
-
-    [HttpGet("{id}")]
-    public async Task<ActionResult<Patient>> GetPatient(int id)
-    {
-        var patient = await _dbContext.Patients.FindAsync(id);
-
-        if (patient == null)
-        {
-            return NotFound();
-        }
-
-        return patient;
-    }
-
-    [HttpPost]
-    public async Task<ActionResult<Patient>> CreatePatient(Patient patient)
-    {
-        _dbContext.Patients.Add(patient);
-        await _dbContext.SaveChangesAsync();
-
-        return CreatedAtAction(nameof(GetPatient), new { id = patient.Id }, patient);
-    }
-
-    [HttpPut("{id}")]
-    public async Task<IActionResult> UpdatePatient(int id, Patient patient)
-    {
-        if (id != patient.Id)
-        {
-            return BadRequest();
-        }
-
-        _dbContext.Entry(patient).State = EntityState.Modified;
-
-        try
-        {
-            await _dbContext.SaveChangesAsync();
-        }
-        catch (DbUpdateConcurrencyException)
-        {
-            if (!PatientExists(id))
-            {
-                return NotFound();
-            }
-            else
-            {
-                throw;
-            }
-        }
-
-        return NoContent();
-    }
-
-    [HttpDelete("{id}")]
-    public async Task<IActionResult> DeletePatient(int id)
-    {
-        var patient = await _dbContext.Patients.FindAsync(id);
-        if (patient == null)
-        {
-            return NotFound();
-        }
-
-        _dbContext.Patients.Remove(patient);
-        await _dbContext.SaveChangesAsync();
-
-        return NoContent();
-    }
-
-    private bool PatientExists(int id)
-    {
-        return _dbContext.Patients.Any(e => e.Id == id);
-    }
-}
