+++ Frontend/Frontend.csproj
@@ -11,4 +11,10 @@
     <ProjectReference Include="..\BackendPatient\BackendPatient.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Models\" />
+    <Folder Include="Views\" />
+    <Folder Include="Controllers\" />
+  </ItemGroup>
+
 </Project>
